8 - simple redux store + prettier
