+++ package.json
@@ -20,6 +20,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "redux": "^4.1.0",
     "typescript": "^4.2.4",
     "web-vitals": "^1.1.1"
   },
