+++ src/store/reducers/app.reducer.ts
@@ -1,0 +1,13 @@
+type AppState = {
+    main: string;
+}
+
+type AppAction = {
+    type: string;
+}
+
+const initialState = {main: "Главная"}
+
+export function appReducer(state: AppState = initialState, action: AppAction): AppState {
+    return state;
+}
