+++ src/store/store.ts
@@ -1,0 +1,6 @@
+import {createStore} from 'redux';
+import {appReducer} from "./reducers/app.reducer";
+
+const store = createStore(appReducer);
+
+export default store;
